requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
cors
cryptography
python-dotenv
